requirements.txt
python==3.12.6
numpy==2.1.1
pydicom==3.0.0
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pypng==0.20220715.0